example13.py

+++ robotarm-python-2021-main/example13.py
@@ -2,23 +2,23 @@ from RobotArm import RobotArm
 
 robotArm = RobotArm('exercise 13')
 robotArm.randomLevel(1,7)
-robotArm.speed = 5
+robotArm.speed = 2
 
 # Jouw python instructies zet je vanaf hier:
-teller = 1
-robotArm.grab()
-for y in range(8):
-    robotArm.moveRight() 
-    robotArm.drop()
-    for x in range(8):    
-        robotArm.moveLeft()
+teller = 9
+bewegen = True
+while bewegen:
     robotArm.grab()
-    for x in range(teller):
-        robotArm.moveRight()
-        teller += 1
-    
- 
-  
+    color = robotArm.scan()
+    if color == "":
+        bewegen = False
+    else:
+        for x in range(0,teller):
+            robotArm.moveRight()
+        robotArm.drop()
+        for x in range(0,teller):
+            robotArm.moveLeft()
+        teller -=1
 
 # Na jouw code wachten tot het sluiten van de window:
 robotArm.wait()
